requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boto3==1.34.34
botocore==1.34.162
celery==5.3.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.4.0
Django==5.2.5
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==37.5.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
reportlab==4.0.8
requests==2.32.5
rpds-py==0.27.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
stripe==8.0.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
